+++ lib/pylib/ci_fw_postfixes.py
@@ -11,6 +11,15 @@ import click
 
 CHECKS = {}
 
+CHECKS['12.0'] = {
+    '.py': [
+        ("replace",
+            r"\.phantom_js\(",
+            ".browser_js(",
+            "Replace calls to 'phantom_js' with 'browser_js'."),
+    ],
+}
+
 # Auto checks for 13.0
 CHECKS['13.0'] = {
     '.py': [
@@ -31,15 +40,24 @@ CHECKS['13.0'] = {
             "._mail_find_partner_from_emails(",
             ("Rename _find_partner_from_emails -> "
                 "_mail_find_partner_from_emails")),
+        ("replace",
+            r"\.phantom_js\(",
+            ".browser_js(",
+            "Replace calls to 'phantom_js' with 'browser_js'."),
     ],
     '.xml': [
         ("replace",
             r"[\s\t]*<field name=['\"]view_type['\"]>.*</field>\n",
             "",
             "Remove <field name='view_type'>...</field>"),
+        ("replace",
+            r"\.phantom_js\(",
+            ".browser_js(",
+            "Replace calls to 'phantom_js' with 'browser_js'."),
     ],
 }
 
+# Auto checks for 14.0
 CHECKS['14.0'] = {
     '.py': [
         ("replace",