+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Short version of `--migrate-modules` option to `odoo-helper ci do-forwardport`: `--mm`.
   So, now forwardport command could look like: `odoo-helper ci do-forward-port -s 12.0 --mm`
+- Automatically detect `.jslintrc` placed in repo root directory. If found, then apply it to pylint by default.
 
 ## Release 0.12.1 (2021-10-25)